move existing unit tests
