+++ src/sinks/kafka-old.rs
@@ -336,80 +336,6 @@ mod tests {
     use std::collections::BTreeMap;
 
     #[test]
-    fn kafka_encode_event_log_text() {
-        crate::test_util::trace_init();
-        let key = "";
-        let message = "hello world".to_string();
-        let (key_bytes, bytes, _metadata) = encode_event(
-            message.clone().into(),
-            &None,
-            &EncodingConfig::from(Encoding::Text),
-        );
-
-        assert_eq!(&key_bytes[..], key.as_bytes());
-        assert_eq!(&bytes[..], message.as_bytes());
-    }
-
-    #[test]
-    fn kafka_encode_event_log_json() {
-        crate::test_util::trace_init();
-        let message = "hello world".to_string();
-        let mut event = Event::from(message.clone());
-        event.as_mut_log().insert("key", "value");
-        event.as_mut_log().insert("foo", "bar");
-
-        let (key, bytes, _metadata) = encode_event(
-            event,
-            &Some("key".into()),
-            &EncodingConfig::from(Encoding::Json),
-        );
-
-        let map: BTreeMap<String, String> = serde_json::from_slice(&bytes[..]).unwrap();
-
-        assert_eq!(&key[..], b"value");
-        assert_eq!(map[&log_schema().message_key().to_string()], message);
-        assert_eq!(map["key"], "value".to_string());
-        assert_eq!(map["foo"], "bar".to_string());
-    }
-
-    #[test]
-    fn kafka_encode_event_metric_text() {
-        let metric = Metric::new(
-            "kafka-metric",
-            MetricKind::Absolute,
-            MetricValue::Counter { value: 0.0 },
-        );
-        let (key_bytes, bytes, _metadata) = encode_event(
-            metric.clone().into(),
-            &None,
-            &EncodingConfig::from(Encoding::Text),
-        );
-
-        assert_eq!("", String::from_utf8_lossy(&key_bytes));
-        assert_eq!(metric.to_string(), String::from_utf8_lossy(&bytes));
-    }
-
-    #[test]
-    fn kafka_encode_event_metric_json() {
-        let metric = Metric::new(
-            "kafka-metric",
-            MetricKind::Absolute,
-            MetricValue::Counter { value: 0.0 },
-        );
-        let (key_bytes, bytes, _metadata) = encode_event(
-            metric.clone().into(),
-            &None,
-            &EncodingConfig::from(Encoding::Json),
-        );
-
-        assert_eq!("", String::from_utf8_lossy(&key_bytes));
-        assert_eq!(
-            serde_json::to_string(&metric).unwrap(),
-            String::from_utf8_lossy(&bytes)
-        );
-    }
-
-    #[test]
     fn kafka_encode_event_log_apply_rules() {
         crate::test_util::trace_init();
         let mut event = Event::from("hello");
@@ -433,355 +359,4 @@ mod tests {
         assert!(!map.contains_key("key"));
     }
 
-    #[test]
-    fn kafka_get_headers() {
-        let headers_key = "headers";
-        let mut header_values = BTreeMap::new();
-        header_values.insert("a-key".to_string(), Value::Bytes(Bytes::from("a-value")));
-        header_values.insert("b-key".to_string(), Value::Bytes(Bytes::from("b-value")));
-
-        let mut event = Event::from("hello");
-        event.as_mut_log().insert(headers_key, header_values);
-
-        let headers = get_headers(&event, headers_key).unwrap();
-        assert_eq!(headers.get(0).unwrap().0, "a-key");
-        assert_eq!(headers.get(0).unwrap().1, "a-value".as_bytes());
-        assert_eq!(headers.get(1).unwrap().0, "b-key");
-        assert_eq!(headers.get(1).unwrap().1, "b-value".as_bytes());
-    }
-}
-
-#[cfg(feature = "kafka-integration-tests")]
-#[cfg(test)]
-mod integration_test {
-    use super::*;
-    use crate::{
-        buffers::Acker,
-        kafka::{KafkaAuthConfig, KafkaSaslConfig, KafkaTlsConfig},
-        test_util::{random_lines_with_stream, random_string, wait_for},
-        tls::TlsOptions,
-    };
-    use bytes::Bytes;
-    use futures::StreamExt;
-    use rdkafka::{
-        consumer::{BaseConsumer, Consumer},
-        message::Headers,
-        Message, Offset, TopicPartitionList,
-    };
-    use std::{collections::BTreeMap, future::ready, thread, time::Duration};
-    use vector_core::event::{BatchNotifier, BatchStatus};
-
-    #[tokio::test]
-    async fn healthcheck() {
-        crate::test_util::trace_init();
-        let topic = format!("test-{}", random_string(10));
-
-        let config = KafkaSinkConfig {
-            bootstrap_servers: "localhost:9091".into(),
-            topic: topic.clone(),
-            key_field: None,
-            encoding: EncodingConfig::from(Encoding::Text),
-            batch: BatchConfig::default(),
-            compression: KafkaCompression::None,
-            auth: KafkaAuthConfig::default(),
-            socket_timeout_ms: 60000,
-            message_timeout_ms: 300000,
-            librdkafka_options: HashMap::new(),
-            headers_key: None,
-        };
-
-        super::healthcheck(config).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_plaintext() {
-        crate::test_util::trace_init();
-        kafka_happy_path("localhost:9091", None, None, KafkaCompression::None).await;
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_gzip() {
-        crate::test_util::trace_init();
-        kafka_happy_path("localhost:9091", None, None, KafkaCompression::Gzip).await;
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_lz4() {
-        crate::test_util::trace_init();
-        kafka_happy_path("localhost:9091", None, None, KafkaCompression::Lz4).await;
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_snappy() {
-        crate::test_util::trace_init();
-        kafka_happy_path("localhost:9091", None, None, KafkaCompression::Snappy).await;
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_zstd() {
-        crate::test_util::trace_init();
-        kafka_happy_path("localhost:9091", None, None, KafkaCompression::Zstd).await;
-    }
-
-    async fn kafka_batch_options_overrides(
-        batch: BatchConfig,
-        librdkafka_options: HashMap<String, String>,
-    ) -> crate::Result<KafkaSink> {
-        let topic = format!("test-{}", random_string(10));
-        let config = KafkaSinkConfig {
-            bootstrap_servers: "localhost:9091".to_string(),
-            topic: format!("{}-%Y%m%d", topic),
-            compression: KafkaCompression::None,
-            encoding: Encoding::Text.into(),
-            key_field: None,
-            auth: KafkaAuthConfig {
-                sasl: None,
-                tls: None,
-            },
-            socket_timeout_ms: 60000,
-            message_timeout_ms: 300000,
-            batch,
-            librdkafka_options,
-            headers_key: None,
-        };
-        let (acker, _ack_counter) = Acker::new_for_testing();
-        config.clone().to_rdkafka(KafkaRole::Consumer)?;
-        config.clone().to_rdkafka(KafkaRole::Producer)?;
-        super::healthcheck(config.clone()).await?;
-        KafkaSink::new(config, acker)
-    }
-
-    #[tokio::test]
-    async fn kafka_batch_options_max_bytes_errors_on_double_set() {
-        crate::test_util::trace_init();
-        assert!(kafka_batch_options_overrides(
-            BatchConfig {
-                max_bytes: Some(1000),
-                max_events: None,
-                max_size: None,
-                timeout_secs: None
-            },
-            indexmap::indexmap! {
-                "batch.size".to_string() => 1.to_string(),
-            }
-            .into_iter()
-            .collect()
-        )
-        .await
-        .is_err())
-    }
-
-    #[tokio::test]
-    async fn kafka_batch_options_actually_sets() {
-        crate::test_util::trace_init();
-        kafka_batch_options_overrides(
-            BatchConfig {
-                max_bytes: None,
-                max_events: Some(10),
-                max_size: None,
-                timeout_secs: Some(2),
-            },
-            indexmap::indexmap! {}.into_iter().collect(),
-        )
-        .await
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn kafka_batch_options_max_events_errors_on_double_set() {
-        crate::test_util::trace_init();
-        assert!(kafka_batch_options_overrides(
-            BatchConfig {
-                max_bytes: None,
-                max_events: Some(10),
-                max_size: None,
-                timeout_secs: None
-            },
-            indexmap::indexmap! {
-                "batch.num.messages".to_string() => 1.to_string(),
-            }
-            .into_iter()
-            .collect()
-        )
-        .await
-        .is_err())
-    }
-
-    #[tokio::test]
-    async fn kafka_batch_options_timeout_secs_errors_on_double_set() {
-        crate::test_util::trace_init();
-        assert!(kafka_batch_options_overrides(
-            BatchConfig {
-                max_bytes: None,
-                max_events: None,
-                max_size: None,
-                timeout_secs: Some(10),
-            },
-            indexmap::indexmap! {
-                "queue.buffering.max.ms".to_string() => 1.to_string(),
-            }
-            .into_iter()
-            .collect()
-        )
-        .await
-        .is_err())
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_tls() {
-        crate::test_util::trace_init();
-        kafka_happy_path(
-            "localhost:9092",
-            None,
-            Some(KafkaTlsConfig {
-                enabled: Some(true),
-                options: TlsOptions::test_options(),
-            }),
-            KafkaCompression::None,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_tls_with_key() {
-        crate::test_util::trace_init();
-        kafka_happy_path(
-            "localhost:9092",
-            None,
-            Some(KafkaTlsConfig {
-                enabled: Some(true),
-                options: TlsOptions::test_options(),
-            }),
-            KafkaCompression::None,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    async fn kafka_happy_path_sasl() {
-        crate::test_util::trace_init();
-        kafka_happy_path(
-            "localhost:9093",
-            Some(KafkaSaslConfig {
-                enabled: Some(true),
-                username: Some("admin".to_owned()),
-                password: Some("admin".to_owned()),
-                mechanism: Some("PLAIN".to_owned()),
-            }),
-            None,
-            KafkaCompression::None,
-        )
-        .await;
-    }
-
-    async fn kafka_happy_path(
-        server: &str,
-        sasl: Option<KafkaSaslConfig>,
-        tls: Option<KafkaTlsConfig>,
-        compression: KafkaCompression,
-    ) {
-        let topic = format!("test-{}", random_string(10));
-        let headers_key = "headers_key".to_string();
-        let kafka_auth = KafkaAuthConfig { sasl, tls };
-        let config = KafkaSinkConfig {
-            bootstrap_servers: server.to_string(),
-            topic: format!("{}-%Y%m%d", topic),
-            key_field: None,
-            encoding: EncodingConfig::from(Encoding::Text),
-            batch: BatchConfig::default(),
-            compression,
-            auth: kafka_auth.clone(),
-            socket_timeout_ms: 60000,
-            message_timeout_ms: 300000,
-            librdkafka_options: HashMap::new(),
-            headers_key: Some(headers_key.clone()),
-        };
-        let topic = format!("{}-{}", topic, chrono::Utc::now().format("%Y%m%d"));
-        let (acker, ack_counter) = Acker::new_for_testing();
-        let sink = KafkaSink::new(config, acker).unwrap();
-
-        let num_events = 1000;
-        let (batch, mut receiver) = BatchNotifier::new_with_receiver();
-        let (input, events) = random_lines_with_stream(100, num_events, Some(batch));
-
-        let header_1_key = "header-1-key";
-        let header_1_value = "header-1-value";
-        events
-            .map(|mut event| {
-                let mut header_values = BTreeMap::new();
-                header_values.insert(
-                    header_1_key.to_string(),
-                    Value::Bytes(Bytes::from(header_1_value)),
-                );
-                event
-                    .as_mut_log()
-                    .insert(headers_key.clone(), header_values);
-                Ok(event)
-            })
-            .forward(sink)
-            .await
-            .unwrap();
-        assert_eq!(receiver.try_recv(), Ok(BatchStatus::Delivered));
-
-        // read back everything from the beginning
-        let mut client_config = rdkafka::ClientConfig::new();
-        client_config.set("bootstrap.servers", server);
-        client_config.set("group.id", &random_string(10));
-        client_config.set("enable.partition.eof", "true");
-        let _ = kafka_auth.apply(&mut client_config).unwrap();
-
-        let mut tpl = TopicPartitionList::new();
-        tpl.add_partition(&topic, 0)
-            .set_offset(Offset::Beginning)
-            .unwrap();
-
-        let consumer: BaseConsumer = client_config.create().unwrap();
-        consumer.assign(&tpl).unwrap();
-
-        // wait for messages to show up
-        wait_for(
-            || match consumer.fetch_watermarks(&topic, 0, Duration::from_secs(3)) {
-                Ok((_low, high)) => ready(high > 0),
-                Err(err) => {
-                    println!("retrying due to error fetching watermarks: {}", err);
-                    ready(false)
-                }
-            },
-        )
-        .await;
-
-        // check we have the expected number of messages in the topic
-        let (low, high) = consumer
-            .fetch_watermarks(&topic, 0, Duration::from_secs(3))
-            .unwrap();
-        assert_eq!((0, num_events as i64), (low, high));
-
-        // loop instead of iter so we can set a timeout
-        let mut failures = 0;
-        let mut out = Vec::new();
-        while failures < 100 {
-            match consumer.poll(Duration::from_secs(3)) {
-                Some(Ok(msg)) => {
-                    let s: &str = msg.payload_view().unwrap().unwrap();
-                    out.push(s.to_owned());
-                    let (header_key, header_val) = msg.headers().unwrap().get(0).unwrap();
-                    assert_eq!(header_key, header_1_key);
-                    assert_eq!(header_val, header_1_value.as_bytes());
-                }
-                None if out.len() >= input.len() => break,
-                _ => {
-                    failures += 1;
-                    thread::sleep(Duration::from_millis(50));
-                }
-            }
-        }
-
-        assert_eq!(out.len(), input.len());
-        assert_eq!(out, input);
-
-        assert_eq!(
-            ack_counter.load(std::sync::atomic::Ordering::Relaxed),
-            num_events
-        );
-    }
 }
