+++ src/sinks/kafka/mod.rs
@@ -5,6 +5,7 @@ pub(crate) mod request_builder;
 pub(crate) mod service;
 pub(crate) mod sink;
 pub(crate) mod tests;
+pub(crate) mod encoder;
 
 // #[cfg(test)]
 // mod tests;
